requirements.txt
python-telegram-bot==13.15
environs==14.1.0
vk-api==11.9.9
